+++ packages/remote/src/electron-node/remote-types.ts
@@ -60,7 +60,7 @@ export interface RemoteConnection extends Disposable {
     /**
      * copy files from local to remote
      */
-    copy(localPath: string | Buffer | NodeJS.ReadableStream, remotePath: string): Promise<void>;
+    copy(localPath: string, remotePath: string): Promise<void>;
 
     /**
      * used for disposing when theia is shutting down